Changed frequency type to DoubleProperty

+++ src/main/java/de/uniks/se19/team_g/project_rbsg/skynet/Bot.java
@@ -1,14 +1,14 @@
 package de.uniks.se19.team_g.project_rbsg.skynet;
 
 import javafx.application.Platform;
-import javafx.beans.property.IntegerProperty;
-import javafx.beans.property.SimpleIntegerProperty;
+import javafx.beans.property.DoubleProperty;
+import javafx.beans.property.SimpleDoubleProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class Bot implements Runnable
 {
-    public IntegerProperty frequency = new SimpleIntegerProperty(2);
+    public DoubleProperty frequency = new SimpleDoubleProperty(2);
     private Skynet skynet;
     private Thread myThread;
     private Logger logger = LoggerFactory.getLogger(getClass());
@@ -47,7 +47,7 @@ public class Bot implements Runnable
             Platform.runLater(() -> skynet.turn());
             try
             {
-                Thread.sleep(1000 / frequency.get());
+                Thread.sleep((int) (1000 / frequency.get()));
             } catch (InterruptedException e)
             {
                 e.printStackTrace();
